// src/app/flows/fines/fines-acc/fines-acc-party-add-amend-convert/interfaces/fines-acc-party-add-amend-convert-organisation-alias-state.interface.ts
export interface IFinesAccPartyAddAmendConvertOrganisationAliasState {
  facc_party_add_amend_convert_alias_organisation_name_0?: string;
  facc_party_add_amend_convert_alias_id_0?: string; // Store original alias ID
  facc_party_add_amend_convert_alias_organisation_name_1?: string;
  facc_party_add_amend_convert_alias_id_1?: string; // Store original alias ID
  facc_party_add_amend_convert_alias_organisation_name_2?: string;
  facc_party_add_amend_convert_alias_id_2?: string; // Store original alias ID
  facc_party_add_amend_convert_alias_organisation_name_3?: string;
  facc_party_add_amend_convert_alias_id_3?: string; // Store original alias ID
  facc_party_add_amend_convert_alias_organisation_name_4?: string;
  facc_party_add_amend_convert_alias_id_4?: string; // Store original alias ID
}